Setting color scheme and text theme
Flutter has different approaches to set the theme, but the most common is to apply them in the way that we are doing right now.
In my opinion we should separate this piece of code in other file or bring them to a persistent data storage, even use it as a key value pair map object in order to be able to create our custom theme.